change the API for SIMD reduction
The user's unary transform functor
is now expected to return a simd
value and will be reduced on the
P3A side to a scalar value.
This will help have more consistent
interfaces between the plain reducer
and the reproducible adder, which will
probably vector store the values to a
temporary array instead of reducing them
(even a simd add reduction is not
associative).