remove p3a::compare and its uses
having "hot code" classes like the mandel
tensors do input validation is inconsistent
with the general design philosophy of P3A.
tensor classes need to be as fast as hand-rolled
code, and unconditionally checking for errors
violates that condition.
Checking correctness should be something the
user can do optionally and not on by default.